SIR Model in one file
Compiling the 4 different python files for the SIR model into one. I belive that the issue with the SIR model not approproatley appearing on the webste has to do bokeh objects not being globally mutable so when each panel was in a different file there was an issue on the server. Compiling all into one file should fix the issue
